+++ packages/egenie-common/src/filterItems/utils.tsx
@@ -1,7 +1,19 @@
 import { Typography } from 'antd';
 import _ from 'lodash';
 import React from 'react';
-import type { ValueAndLabelData } from './types';
+import { FilterCascader } from './filterCascader';
+import { FilterCheckbox } from './filterCheckbox';
+import { FilterDate } from './filterDate';
+import { FilterDateStartOrEnd } from './filterDateStartOrEnd';
+import { FilterInput } from './filterInput';
+import { FilterInputAndSelect } from './filterInputAndSelect';
+import { FilterInputNumberGroup } from './filterInputNumberGroup';
+import { FilterInputOrSelect } from './filterInputOrSelect';
+import { FilterRadio } from './filterRadio';
+import { FilterSelect } from './filterSelect';
+import { FilterTreeSelect } from './filterTreeSelect';
+import type { FilterItem, FilterItemOptions, ValueAndLabelData } from './types';
+import { ENUM_FILTER_ITEM_TYPE } from './types';
 
 export function formatValueAndLabelData(data: ValueAndLabelData): ValueAndLabelData {
   if (Array.isArray(data)) {
@@ -65,3 +77,34 @@ export function trimWhiteSpace(value: string, isTrimWhiteSpace: boolean): string
 }
 
 export const throttleTime = 300;
+
+// @ts-ignore
+export function filterInstanceFactory(item: FilterItemOptions): FilterItem {
+  switch (item.type) {
+    case ENUM_FILTER_ITEM_TYPE.input:
+      return new FilterInput(item);
+    case ENUM_FILTER_ITEM_TYPE.inputNumberGroup:
+      return new FilterInputNumberGroup(item);
+    case ENUM_FILTER_ITEM_TYPE.select:
+      return new FilterSelect(item);
+    case ENUM_FILTER_ITEM_TYPE.radio:
+      return new FilterRadio(item);
+    case ENUM_FILTER_ITEM_TYPE.inputAndSelect:
+      return new FilterInputAndSelect(item);
+    case ENUM_FILTER_ITEM_TYPE.date:
+    case ENUM_FILTER_ITEM_TYPE.dateRange:
+      return new FilterDate(item);
+    case ENUM_FILTER_ITEM_TYPE.dateStart:
+    case ENUM_FILTER_ITEM_TYPE.dateEnd:
+      return new FilterDateStartOrEnd(item);
+    case ENUM_FILTER_ITEM_TYPE.checkbox:
+      return new FilterCheckbox(item);
+    case ENUM_FILTER_ITEM_TYPE.inputOrSelect:
+      return new FilterInputOrSelect(item);
+    case ENUM_FILTER_ITEM_TYPE.cascader:
+      return new FilterCascader(item);
+    case ENUM_FILTER_ITEM_TYPE.treeSelect:
+      return new FilterTreeSelect(item);
+  }
+}
+